Split knowhere.h file
