+++ include/knowhere/index.h
@@ -1,98 +1,10 @@
-#ifndef KNOWHERE_H
-#define KNOWHERE_H
+#ifndef INDEX_H
+#define INDEX_H
 
-#include <atomic>
-#include <functional>
-#include <iostream>
-#include <unordered_map>
-
-#include "knowhere/binaryset.h"
-#include "knowhere/bitsetview.h"
-#include "knowhere/config.h"
-#include "knowhere/dataset.h"
-#include "knowhere/file_manager.h"
+#include "knowhere/index_node.h"
 #include "knowhere/log.h"
+
 namespace knowhere {
-
-class Object {
- public:
-    Object() = default;
-    Object(const std::nullptr_t value) {
-        assert(value == nullptr);
-    }
-    inline uint32_t
-    Ref() const {
-        return ref_counts_.load(std::memory_order_relaxed);
-    }
-    inline void
-    DecRef() {
-        ref_counts_.fetch_sub(1, std::memory_order_relaxed);
-    }
-    inline void
-    IncRef() {
-        ref_counts_.fetch_add(1, std::memory_order_relaxed);
-    }
-    virtual ~Object() {
-    }
-
- private:
-    mutable std::atomic_uint32_t ref_counts_ = 1;
-};
-
-template <typename T>
-class Pack : public Object {
-    static_assert(std::is_same_v<T, std::shared_ptr<knowhere::FileManager>>,
-                  "IndexPack only support std::shared_ptr<knowhere::FileManager> by far.");
-
- public:
-    Pack() {
-    }
-    Pack(T package) : package_(package) {
-    }
-    T
-    GetPack() const {
-        return package_;
-    }
-    ~Pack() {
-    }
-
- private:
-    T package_;
-};
-
-class IndexNode : public Object {
- public:
-    virtual Status
-    Build(const DataSet& dataset, const Config& cfg) = 0;
-    virtual Status
-    Train(const DataSet& dataset, const Config& cfg) = 0;
-    virtual Status
-    Add(const DataSet& dataset, const Config& cfg) = 0;
-    virtual expected<DataSetPtr, Status>
-    Search(const DataSet& dataset, const Config& cfg, const BitsetView& bitset) const = 0;
-    virtual expected<DataSetPtr, Status>
-    RangeSearch(const DataSet& dataset, const Config& cfg, const BitsetView& bitset) const = 0;
-    virtual expected<DataSetPtr, Status>
-    GetVectorByIds(const DataSet& dataset, const Config& cfg) const = 0;
-    virtual expected<DataSetPtr, Status>
-    GetIndexMeta(const Config& cfg) const = 0;
-    virtual Status
-    Serialize(BinarySet& binset) const = 0;
-    virtual Status
-    Deserialize(const BinarySet& binset) = 0;
-    virtual std::unique_ptr<BaseConfig>
-    CreateConfig() const = 0;
-    virtual int64_t
-    Dim() const = 0;
-    virtual int64_t
-    Size() const = 0;
-    virtual int64_t
-    Count() const = 0;
-    virtual std::string
-    Type() const = 0;
-    virtual ~IndexNode() {
-    }
-};
 
 template <typename T1>
 class Index {
@@ -308,26 +220,6 @@ class Index {
     T1* node;
 };
 
-class IndexFactory {
- public:
-    Index<IndexNode>
-    Create(const std::string& name, const Object& object = nullptr);
-    const IndexFactory&
-    Register(const std::string& name, std::function<Index<IndexNode>(const Object&)> func);
-    static IndexFactory&
-    Instance();
-
- private:
-    typedef std::map<std::string, std::function<Index<IndexNode>(const Object&)>> FuncMap;
-    IndexFactory();
-    static FuncMap&
-    MapInstance();
-};
-
-#define KNOWHERE_CONCAT(x, y) x##y
-#define KNOWHERE_REGISTER_GLOBAL(name, func) \
-    const IndexFactory& KNOWHERE_CONCAT(index_factory_ref_, name) = IndexFactory::Instance().Register(#name, func)
-
 }  // namespace knowhere
 
-#endif /* KNOWHERE_H */
+#endif /* INDEX_H */
