+++ src/index/ivf_gpu/ivf_gpu.cc
@@ -10,7 +10,7 @@
 #include "faiss/index_io.h"
 #include "index/ivf_gpu/ivf_gpu_config.h"
 #include "io/FaissIO.h"
-#include "knowhere/knowhere.h"
+#include "knowhere/factory.h"
 
 namespace knowhere {
